fix(esbuild): roll back original container mapping system
