+++ apps/esbuild/mfe1/main.ts
@@ -1,6 +1,6 @@
-import { initFederation } from '@module-federation/esbuild';
+// import { initFederation } from '@module-federation/esbuild';
 
 (async () => {
-  await initFederation(undefined, { name: 'mfe1' });
+  // await initFederation(undefined, { name: 'mfe1' });
   await import('./bootstrap');
 })();
